intégration image 5
modification calcule similarité image

+++ package_image/comparaison_image.h
@@ -11,7 +11,7 @@ struct info_comparaison
 {
     char id_image[200];
     int difference[64];
-    double somme;
+    float somme;
     int valeur_token_indexe[64];
     char couleur_indexe[10];
     float taux_de_similarite;
